[#103] fix: 티어 상세 정보 constants로 분리, 티어 이미지가 부모의 크기에 맞춰지도록 변경
