+++ client/src/constants/tierInfo.ts
@@ -1,0 +1,38 @@
+interface TierInfoProps {
+  level: 0 | 1 | 2 | 3 | 4 | 5;
+  text: string;
+  label: string;
+}
+
+export const tierInfo: TierInfoProps[] = [
+  {
+    level: 0,
+    text: 'GUEST',
+    label: '깃털',
+  },
+  {
+    level: 1,
+    text: 'Lv.1',
+    label: '달걀',
+  },
+  {
+    level: 2,
+    text: 'Lv.2',
+    label: '갓 난 병아리',
+  },
+  {
+    level: 3,
+    text: 'Lv.3',
+    label: '병아리',
+  },
+  {
+    level: 4,
+    text: 'Lv.4',
+    label: '청소년 병아리',
+  },
+  {
+    level: 5,
+    text: 'Lv.5',
+    label: '위풍당당한 닭',
+  },
+];
